Generalize legacy skin player test.
