+++ osu.Game/Tests/Visual/LegacySkinPlayerTestScene.cs
@@ -1,0 +1,38 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using NUnit.Framework;
+using osu.Framework.Allocation;
+using osu.Framework.Audio;
+using osu.Framework.IO.Stores;
+using osu.Game.Rulesets;
+using osu.Game.Skinning;
+
+namespace osu.Game.Tests.Visual
+{
+    [TestFixture]
+    public abstract class LegacySkinPlayerTestScene : PlayerTestScene
+    {
+        private ISkinSource legacySkinSource;
+
+        protected override TestPlayer CreatePlayer(Ruleset ruleset) => new SkinProvidingPlayer(legacySkinSource);
+
+        [BackgroundDependencyLoader]
+        private void load(AudioManager audio, OsuGameBase game)
+        {
+            var legacySkin = new DefaultLegacySkin(new NamespacedResourceStore<byte[]>(game.Resources, "Skins/Legacy"), audio);
+            legacySkinSource = new SkinProvidingContainer(legacySkin);
+        }
+
+        public class SkinProvidingPlayer : TestPlayer
+        {
+            [Cached(typeof(ISkinSource))]
+            private readonly ISkinSource skinSource;
+
+            public SkinProvidingPlayer(ISkinSource skinSource)
+            {
+                this.skinSource = skinSource;
+            }
+        }
+    }
+}
